- 백엔드 : 컨트롤러, 서비스, 레포지토리, 엔티티 생성

+++ BACK/src/main/java/com/project/cooperation/model/Member.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.IdGeneratorType;
 import org.hibernate.annotations.Type;
-import org.springframework.data.annotation.Id;
 
 @Entity
 @Table(name = "member")
@@ -17,7 +16,7 @@ public class Member {
     @Id
     @Column(name = "member_idx")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idx;
+    private Long idx;
 
     @Column(name = "member_id")
     private String id;
